feat:서비스 코드 이동 및 리펙토링
