+++ cores/NetCore.h
@@ -108,6 +108,7 @@ size_t get_rear_send_buf_size(void_queue_t* vq);
 void reserve_send(void_queue_t* vq, char* send_org, int send_size);
 // ✨ 서비스 함수. 이런 형태의 함수들을 추가하여 서비스 추가. ✨
 void echo_service(epoll_net_core* server_ptr, task_t* task);
+void login_service(epoll_net_core* server_ptr, task_t* task);
 void signup_service(epoll_net_core* server_ptr, task_t* task);
 void make_group_service(epoll_net_core* server_ptr, task_t* task);
 void user_list_service(epoll_net_core* server_ptr, task_t* task);
