+++ mariadb/mariadb.h
@@ -25,7 +25,9 @@ typedef struct chatdb {
     int db_sizes[TOTAL_DB_NUM];
 } chatdb_t;
 
+bool has_query_results(conn_t* conn, char** msg, const char* query);
+cJSON* execute_query(conn_t* conn, char** msg, const char* query, int key_num, ...);
+void release_conns(chatdb_t* db, int release_conn_num, ...);
 bool init_mariadb(chatdb_t* db);
 void close_mariadb(chatdb_t* db);
-
 #endif
